First commit. Working implementation of the challenge on the instagram blog.
This is a solution to the challenge on the instagram blog: http://instagram-engineering.tumblr.com/post/12651721845/instagram-engineering-challenge-the-unshredder#notes

It uses a totally arbitrary threshold of +- 10 for each RGB color value to compare pixels -- if all three values are within 10 of each other, the pixels are considered similar, otherwise they are different.

It matches up the shreds based on first finding the leftmost shred and then iteratively pasting in the most similar shred to the right, and then repeating with that new shred as the left shred.